refactor: rename ApiService to PersonalIdApiService and move to connectId package
